fix: minor bugs and issues

+++ api/app/Modules/Production/Services/WorkOrderService.php
@@ -294,7 +294,7 @@ class WorkOrderService
                 $end = Carbon::now();
                 $open->update([
                     'end_time'         => $end,
-                    'duration_minutes' => max(0, $open->start_time->diffInMinutes($end)),
+                    'duration_minutes' => (int) max(0, $open->start_time->diffInMinutes($end)),
                 ]);
             }
             if ($wo->machine_id) {
